requirements.txt
pandas
django
djangorestframework
numpy
joblib
scikit-learn
matplotlib
pydantic